+++ src/FBN/GuideBundle/DataFixtures/ORM/ImageRestaurant.php
@@ -29,12 +29,17 @@ class ImageRestaurant extends AbstractFixture implements OrderedFixtureInterface
         $path = __DIR__.'/Resources/Images/restaurants/';
         $pathto = __DIR__.'/../../../../../web/uploads/images/restaurants/';
 
-        // First delete all files in target directory
-        $files = glob($pathto.'/*');
-        foreach ($files as $file) {
-            if (is_file($file)) {
-                unlink($file);
+        // Empty target directory if it exists
+        if (file_exists($pathto)) {
+            $files = scandir($pathto);
+            foreach ($files as $file) {
+                if (is_file($file)) {
+                    unlink($file);
+                }
             }
+        // Create target directory
+        } else {
+            mkdir($pathto, 0777, true);
         }
 
         // Clean cache (whole cache as this file is the first loaded for images fixtures)
